requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
chardet==3.0.4
Cython==0.29.32
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==2.8
joblib==0.14.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
psycopg2==2.8.2
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.25.11
Werkzeug==2.2.2
wrapt==1.14.1